18th commit by yangmingsen
内容：
     1.完成300类型数据处理

author:yms
date: 2020-07-06

+++ galaplat-comprehensive-bidding-service/src/main/java/com/galaplat/comprehensive/bidding/netty/pojo/res/Res300t2.java
@@ -1,14 +1,16 @@
 package com.galaplat.comprehensive.bidding.netty.pojo.res;
 
+import java.math.BigDecimal;
+
 public class Res300t2 {
-    private Integer bid;
+    private BigDecimal bid;
     private String bidTime;
 
-    public Integer getBid() {
+    public BigDecimal getBid() {
         return bid;
     }
 
-    public void setBid(Integer bid) {
+    public void setBid(BigDecimal bid) {
         this.bid = bid;
     }
